test(runtime): isolate the tracing-capture test in its own binary
